git-svn-id: file:///home/svn/opt_utilities@4 ed2142bd-67ad-457f-ba7c-d818d4011675

+++ core/default_data_set.hpp
@@ -20,12 +20,12 @@ private:
   }
 
 
-  const data<Ty,Tx>& do_get_data(int i)const
+  const data<Ty,Tx>& do_get_data(size_t i)const
   {
     return data_vec.at(i);
   }
   
-  int do_size()const
+  size_t do_size()const
   {
     return data_vec.size();
   }
